All column cast functionality moved to cast.cc
