+++ c/column.cc
@@ -222,79 +222,6 @@ size_t Column::nmodal() const  { return get_stats()->nmodal(this); }
 
 
 
-//------------------------------------------------------------------------------
-// Casting
-//------------------------------------------------------------------------------
-
-Column* Column::cast(SType new_stype) const {
-  return cast(new_stype, MemoryRange());
-}
-
-Column* Column::cast(SType new_stype, MemoryRange&& mr) const {
-  if (ri) {
-    // TODO: implement this
-    throw RuntimeError() << "Cannot cast a column with rowindex";
-  }
-  Column *res = nullptr;
-  if (mr) {
-    res = Column::new_column(new_stype);
-    res->nrows = nrows;
-    res->mbuf = std::move(mr);
-  } else {
-    if (new_stype == stype()) {
-      return shallowcopy();
-    }
-    res = Column::new_data_column(new_stype, nrows);
-  }
-  switch (new_stype) {
-    case SType::BOOL:    cast_into(static_cast<BoolColumn*>(res)); break;
-    case SType::INT8:    cast_into(static_cast<IntColumn<int8_t>*>(res)); break;
-    case SType::INT16:   cast_into(static_cast<IntColumn<int16_t>*>(res)); break;
-    case SType::INT32:   cast_into(static_cast<IntColumn<int32_t>*>(res)); break;
-    case SType::INT64:   cast_into(static_cast<IntColumn<int64_t>*>(res)); break;
-    case SType::FLOAT32: cast_into(static_cast<RealColumn<float>*>(res)); break;
-    case SType::FLOAT64: cast_into(static_cast<RealColumn<double>*>(res)); break;
-    case SType::STR32:   cast_into(static_cast<StringColumn<uint32_t>*>(res)); break;
-    case SType::STR64:   cast_into(static_cast<StringColumn<uint64_t>*>(res)); break;
-    case SType::OBJ:     cast_into(static_cast<PyObjectColumn*>(res)); break;
-    default:
-      throw ValueError() << "Unable to cast into stype = " << new_stype;
-  }
-  return res;
-}
-
-void Column::cast_into(BoolColumn*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into bool";
-}
-void Column::cast_into(IntColumn<int8_t>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into int8";
-}
-void Column::cast_into(IntColumn<int16_t>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into int16";
-}
-void Column::cast_into(IntColumn<int32_t>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into int32";
-}
-void Column::cast_into(IntColumn<int64_t>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into int64";
-}
-void Column::cast_into(RealColumn<float>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into float";
-}
-void Column::cast_into(RealColumn<double>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into double";
-}
-void Column::cast_into(StringColumn<uint32_t>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into str32";
-}
-void Column::cast_into(StringColumn<uint64_t>*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into str64";
-}
-void Column::cast_into(PyObjectColumn*) const {
-  throw ValueError() << "Cannot cast " << stype() << " into pyobj";
-}
-
-
 
 //------------------------------------------------------------------------------
 // Integrity checks
