+++ c/column_bool.cc
@@ -44,103 +44,6 @@ double  BoolColumn::sd() const   { return get_stats()->stdev(this); }
 
 
 
-//------------------------------------------------------------------------------
-// Type casts
-//------------------------------------------------------------------------------
-typedef std::unique_ptr<MemoryWritableBuffer> MWBPtr;
-
-template <typename T>
-inline static void cast_helper(size_t nrows, const int8_t* src, T* trg) {
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < nrows; ++i) {
-    int8_t x = src[i];
-    trg[i] = ISNA<int8_t>(x)? GETNA<T>() : static_cast<T>(x);
-  }
-}
-
-template <typename T>
-inline static MemoryRange cast_str_helper(
-  const BoolColumn* src, StringColumn<T>* target)
-{
-  const int8_t* src_data = src->elements_r();
-  T* toffsets = target->offsets_w();
-  size_t exp_size = src->nrows;
-  auto wb = MWBPtr(new MemoryWritableBuffer(exp_size));
-  char* tmpbuf = new char[1024];
-  TRACK(tmpbuf, sizeof(tmpbuf), "BoolColumn::tmpbuf");
-  char* tmpend = tmpbuf + 1000;  // Leave at least 24 spare chars in buffer
-  char* ch = tmpbuf;
-  T offset = 0;
-  toffsets[-1] = 0;
-  for (size_t i = 0; i < src->nrows; ++i) {
-    int8_t x = src_data[i];
-    if (ISNA<int8_t>(x)) {
-      toffsets[i] = offset | GETNA<T>();
-    } else {
-      *ch++ = '0' + x;
-      offset++;
-      toffsets[i] = offset;
-      if (ch > tmpend) {
-        wb->write(static_cast<size_t>(ch - tmpbuf), tmpbuf);
-        ch = tmpbuf;
-      }
-    }
-  }
-  wb->write(static_cast<size_t>(ch - tmpbuf), tmpbuf);
-  wb->finalize();
-  delete[] tmpbuf;
-  UNTRACK(tmpbuf);
-  return wb->get_mbuf();
-}
-
-
-void BoolColumn::cast_into(BoolColumn* target) const {
-  std::memcpy(target->data_w(), data(), alloc_size());
-}
-
-void BoolColumn::cast_into(IntColumn<int8_t>* target) const {
-  std::memcpy(target->data_w(), data(), alloc_size());
-}
-
-void BoolColumn::cast_into(IntColumn<int16_t>* target) const {
-  cast_helper<int16_t>(nrows, this->elements_r(), target->elements_w());
-}
-
-void BoolColumn::cast_into(IntColumn<int32_t>* target) const {
-  cast_helper<int32_t>(nrows, this->elements_r(), target->elements_w());
-}
-
-void BoolColumn::cast_into(IntColumn<int64_t>* target) const {
-  cast_helper<int64_t>(nrows, this->elements_r(), target->elements_w());
-}
-
-void BoolColumn::cast_into(RealColumn<float>* target) const {
-  cast_helper<float>(nrows, this->elements_r(), target->elements_w());
-}
-
-void BoolColumn::cast_into(RealColumn<double>* target) const {
-  cast_helper<double>(nrows, this->elements_r(), target->elements_w());
-}
-
-void BoolColumn::cast_into(PyObjectColumn* target) const {
-  const int8_t* src_data = this->elements_r();
-  PyObject**    trg_data = target->elements_w();
-  for (size_t i = 0; i < nrows; ++i) {
-    trg_data[i] = py::obool(src_data[i]).release();
-  }
-}
-
-void BoolColumn::cast_into(StringColumn<uint32_t>* target) const {
-  MemoryRange strbuf = cast_str_helper<uint32_t>(this, target);
-  target->replace_buffer(target->data_buf(), std::move(strbuf));
-}
-
-void BoolColumn::cast_into(StringColumn<uint64_t>* target) const {
-  MemoryRange strbuf = cast_str_helper<uint64_t>(this, target);
-  target->replace_buffer(target->data_buf(), std::move(strbuf));
-}
-
-
 
 //------------------------------------------------------------------------------
 // Integrity checks
