+++ c/column_int.cc
@@ -6,7 +6,6 @@
 // © H2O.ai 2018
 //------------------------------------------------------------------------------
 #include "column.h"
-#include "csv/toa.h"
 #include "datatablemodule.h"
 #include "python/int.h"
 #include "py_types.h"
@@ -62,153 +61,6 @@ template <typename T>
 int64_t IntColumn<T>::max_int64() const {
   T x = max();
   return ISNA<T>(x)? GETNA<int64_t>() : static_cast<int64_t>(x);
-}
-
-
-
-//------------------------------------------------------------------------------
-// Type casts
-//------------------------------------------------------------------------------
-typedef std::unique_ptr<MemoryWritableBuffer> MWBPtr;
-
-template<typename IT, typename OT>
-inline static void cast_helper(size_t nrows, const IT* src, OT* trg) {
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < nrows; ++i) {
-    IT x = src[i];
-    trg[i] = ISNA<IT>(x)? GETNA<OT>() : static_cast<OT>(x);
-  }
-}
-
-template<typename IT, typename OT>
-inline static MemoryRange cast_str_helper(
-  size_t nrows, const IT* src, OT* toffsets)
-{
-  size_t exp_size = nrows * sizeof(IT);
-  auto wb = MWBPtr(new MemoryWritableBuffer(exp_size));
-  char* tmpbuf = new char[1024];
-  TRACK(tmpbuf, sizeof(tmpbuf), "IntColumn::tmpbuf");
-  char* tmpend = tmpbuf + 1000;  // Leave at least 24 spare chars in buffer
-  char* ch = tmpbuf;
-  OT offset = 0;
-  toffsets[-1] = 0;
-  for (size_t i = 0; i < nrows; ++i) {
-    IT x = src[i];
-    if (ISNA<IT>(x)) {
-      toffsets[i] = offset | GETNA<OT>();
-    } else {
-      char* ch0 = ch;
-      toa<IT>(&ch, x);
-      offset += static_cast<OT>(ch - ch0);
-      toffsets[i] = offset;
-      if (ch > tmpend) {
-        wb->write(static_cast<size_t>(ch - tmpbuf), tmpbuf);
-        ch = tmpbuf;
-      }
-    }
-  }
-  wb->write(static_cast<size_t>(ch - tmpbuf), tmpbuf);
-  wb->finalize();
-  delete[] tmpbuf;
-  UNTRACK(tmpbuf);
-  return wb->get_mbuf();
-}
-
-
-template <typename T>
-void IntColumn<T>::cast_into(BoolColumn* target) const {
-  constexpr T na_src = GETNA<T>();
-  constexpr int8_t na_trg = GETNA<int8_t>();
-  const T* src_data = this->elements_r();
-  int8_t* trg_data = target->elements_w();
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < this->nrows; ++i) {
-    T x = src_data[i];
-    trg_data[i] = x == na_src? na_trg : (x != 0);
-  }
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(IntColumn<int8_t>* target) const {
-  cast_helper<T, int8_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(IntColumn<int16_t>* target) const {
-  cast_helper<T, int16_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(IntColumn<int32_t>* target) const {
-  cast_helper<T, int32_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(IntColumn<int64_t>* target) const {
-  cast_helper<T, int64_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(RealColumn<float>* target) const {
-  cast_helper<T, float>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(RealColumn<double>* target) const {
-  cast_helper<T, double>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(StringColumn<uint32_t>* target) const {
-  uint32_t* offsets = target->offsets_w();
-  MemoryRange strbuf = cast_str_helper<T, uint32_t>(
-      this->nrows, this->elements_r(), offsets
-  );
-  target->replace_buffer(target->data_buf(), std::move(strbuf));
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(StringColumn<uint64_t>* target) const {
-  uint64_t* offsets = target->offsets_w();
-  MemoryRange strbuf = cast_str_helper<T, uint64_t>(
-      this->nrows, this->elements_r(), offsets
-  );
-  target->replace_buffer(target->data_buf(), std::move(strbuf));
-}
-
-template <typename T>
-void IntColumn<T>::cast_into(PyObjectColumn* target) const {
-  constexpr T na_src = GETNA<T>();
-  const T* src_data = this->elements_r();
-  PyObject** trg_data = target->elements_w();
-  for (size_t i = 0; i < this->nrows; ++i) {
-    T x = src_data[i];
-    // PyLong_FromInt64 is declared in "py_types.h" as an alias for either
-    // PyLong_FromLong or PyLong_FromLongLong, depending on the platform
-    trg_data[i] = x == na_src ? none() :
-                  (sizeof(T) == 8? PyLong_FromInt64(x) :
-                                   PyLong_FromLong(static_cast<long>(x)));
-  }
-}
-
-template <>
-void IntColumn<int8_t>::cast_into(IntColumn<int8_t>* target) const {
-  std::memcpy(target->data_w(), this->data(), alloc_size());
-}
-
-template <>
-void IntColumn<int16_t>::cast_into(IntColumn<int16_t>* target) const {
-  std::memcpy(target->data_w(), this->data(), alloc_size());
-}
-
-template <>
-void IntColumn<int32_t>::cast_into(IntColumn<int32_t>* target) const {
-  std::memcpy(target->data_w(), this->data(), alloc_size());
-}
-
-template <>
-void IntColumn<int64_t>::cast_into(IntColumn<int64_t>* target) const {
-  std::memcpy(target->data_w(), this->data(), alloc_size());
 }
