+++ c/column_real.cc
@@ -50,143 +50,6 @@ template <typename T> double RealColumn<T>::kurt() const { return get_stats()->k
 
 
 
-//------------------------------------------------------------------------------
-// Type casts
-//------------------------------------------------------------------------------
-typedef std::unique_ptr<MemoryWritableBuffer> MWBPtr;
-
-template<typename IT, typename OT>
-inline static void cast_helper(size_t nrows, const IT* src, OT* trg) {
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < nrows; ++i) {
-    IT x = src[i];
-    trg[i] = ISNA<IT>(x)? GETNA<OT>() : static_cast<OT>(x);
-  }
-}
-
-template<typename IT, typename OT>
-inline static MemoryRange cast_str_helper(
-  const RealColumn<IT>* src, StringColumn<OT>* target)
-{
-  const IT* src_data = src->elements_r();
-  OT* toffsets = target->offsets_w();
-  size_t exp_size = src->nrows * sizeof(IT) * 2;
-  auto wb = MWBPtr(new MemoryWritableBuffer(exp_size));
-  char* tmpbuf = new char[1024];
-  TRACK(tmpbuf, sizeof(tmpbuf), "RealColumn::tmpbuf");
-  char* tmpend = tmpbuf + 1000;  // Leave at least 24 spare chars in buffer
-  char* ch = tmpbuf;
-  OT offset = 0;
-  toffsets[-1] = 0;
-  for (size_t i = 0; i < src->nrows; ++i) {
-    IT x = src_data[i];
-    if (ISNA<IT>(x)) {
-      toffsets[i] = offset | GETNA<OT>();
-    } else {
-      char* ch0 = ch;
-      toa<IT>(&ch, x);
-      offset += static_cast<OT>(ch - ch0);
-      toffsets[i] = offset;
-      if (ch > tmpend) {
-        wb->write(static_cast<size_t>(ch - tmpbuf), tmpbuf);
-        ch = tmpbuf;
-      }
-    }
-  }
-  wb->write(static_cast<size_t>(ch - tmpbuf), tmpbuf);
-  wb->finalize();
-  delete[] tmpbuf;
-  UNTRACK(tmpbuf);
-  return wb->get_mbuf();
-}
-
-
-template <typename T>
-void RealColumn<T>::cast_into(BoolColumn* target) const {
-  constexpr int8_t na_trg = GETNA<int8_t>();
-  const T* src_data = this->elements_r();
-  int8_t* trg_data = target->elements_w();
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < this->nrows; ++i) {
-    T x = src_data[i];
-    trg_data[i] = ISNA<T>(x)? na_trg : (x != 0);
-  }
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(IntColumn<int8_t>* target) const {
-  cast_helper<T, int8_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(IntColumn<int16_t>* target) const {
-  cast_helper<T, int16_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(IntColumn<int32_t>* target) const {
-  cast_helper<T, int32_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(IntColumn<int64_t>* target) const {
-  cast_helper<T, int64_t>(this->nrows, this->elements_r(), target->elements_w());
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(StringColumn<uint32_t>* target) const {
-  MemoryRange strbuf = cast_str_helper<T, uint32_t>(this, target);
-  target->replace_buffer(target->data_buf(), std::move(strbuf));
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(StringColumn<uint64_t>* target) const {
-  MemoryRange strbuf = cast_str_helper<T, uint64_t>(this, target);
-  target->replace_buffer(target->data_buf(), std::move(strbuf));
-}
-
-template <>
-void RealColumn<float>::cast_into(RealColumn<double>* target) const {
-  const float* src_data = this->elements_r();
-  double* trg_data = target->elements_w();
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < this->nrows; ++i) {
-    trg_data[i] = static_cast<double>(src_data[i]);
-  }
-}
-
-template <>
-void RealColumn<double>::cast_into(RealColumn<float>* target) const {
-  const double* src_data = this->elements_r();
-  float* trg_data = target->elements_w();
-  #pragma omp parallel for schedule(static)
-  for (size_t i = 0; i < this->nrows; ++i) {
-    trg_data[i] = static_cast<float>(src_data[i]);
-  }
-}
-
-template <>
-void RealColumn<float>::cast_into(RealColumn<float>* target) const {
-  std::memcpy(target->data_w(), this->data(), alloc_size());
-}
-
-template <>
-void RealColumn<double>::cast_into(RealColumn<double>* target) const {
-  std::memcpy(target->data_w(), this->data(), alloc_size());
-}
-
-template <typename T>
-void RealColumn<T>::cast_into(PyObjectColumn* target) const {
-  const T* src_data = this->elements_r();
-  PyObject** trg_data = target->elements_w();
-  for (size_t i = 0; i < this->nrows; ++i) {
-    T x = src_data[i];
-    trg_data[i] = ISNA<T>(x)? none()
-                            : PyFloat_FromDouble(static_cast<double>(x));
-  }
-}
-
-
 
 
 //------------------------------------------------------------------------------
